feat: move beacon beam to common
