+++ 1.21/src/main/java/io/github/axolotlclient/AxolotlClient.java
@@ -35,7 +35,7 @@ import io.github.axolotlclient.modules.freelook.Freelook;
 import io.github.axolotlclient.modules.hud.HudManager;
 import io.github.axolotlclient.modules.hypixel.HypixelMods;
 import io.github.axolotlclient.modules.particles.Particles;
-import io.github.axolotlclient.modules.renderOptions.BeaconBeam;
+import io.github.axolotlclient.modules.render.BeaconBeam;
 import io.github.axolotlclient.modules.rpc.DiscordRPC;
 import io.github.axolotlclient.modules.screenshotUtils.ScreenshotUtils;
 import io.github.axolotlclient.modules.scrollableTooltips.ScrollableTooltips;
@@ -67,7 +67,6 @@ public class AxolotlClient extends AxolotlClientCommon implements ClientModIniti
 		registerModule(TntTime.getInstance());
 		registerModule(Particles.getInstance());
 		registerModule(ScreenshotUtils.getInstance());
-		registerModule(BeaconBeam.getInstance());
 		registerModule(Tablist.getInstance());
 		registerModule(Auth.getInstance());
 		registerModule(APIOptions.getInstance());
